[FIX] cart type 변경, 헤더에서 입력받을 수 있는  X-User-Id로 변경

+++ src/main/java/com/team1/goorm/domain/dto/CartResponseDto.java
@@ -15,7 +15,7 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @Builder
 public class CartResponseDto {
-    private int cartId;
+    private Long cartId;
     private String productName; // 상품 이름
     private BigDecimal price; // 상품 가격
     private int quantity; // 상품 갯수
